initial.rev99
Goal of this branch is to also make it so that poll'ing is stopped when
no paths are there. As polling while no paths being watched is useless
overhead.

COMMIT TO SAVE
this commit shows that i dont have to keep re-.ptr'ing to get the
contents, and i can change the contents from wherever

### bootstrap.js
* Moving `waitForNextChange` out of prototype as this is just a helper
function for the mainthread api, it should not be exposed to devuser's
* Moved the os specific start poll out of `do_createWatcher`