FSEvents api udnerstanding
i changed from macStuff.cId  to
ostypes.CONST.kFSEventStreamEventIdSinceNow when creating new stream.
cuz when devuser adds a path they want events from the time they added
it. not from the time the watcher was created, and with macStuff.cId
when user later adds a path it is triggering the old stuff too.

so change made testing now