sub design decision
made so that watcherid is not of main concern. makes things very clear.
and watch is now global then by watcherid. so multipler watchers
watching the same path does not cause an extra watch to happen, they
will just get the callback triggered.

not specifically due to this sub decision decision, but due to the way i
made it, no limit, as threads are spawned when thread limit is reached
for active.

design adjustment tested on windows its working (for what i had so far).
gio not yet tested.