win and design for removePath and close and shutdown
if watching only signle path it works as expected. but with multi path
it needs more testing.

designed what should happen on removePath and close.

pollers are being destroyed.

on win, the poller is destroyed however it uses a setTimeout of 1000,
TODO: make it so that i dont need to wait, i need to figure out when it
is safe to terminate the worker

can uninstall the addon and it will terminate thigns properly so it wont
have a poll going on when it gets terminated